+++ src/sc/embedder/config.py
@@ -1,2 +1,3 @@
 # -*- coding: utf-8 -*-
 PROJECTNAME = 'sc.embedder'
+PROFILE = PROJECTNAME + ':default'