+++ src/sc/embedder/tests/test_upgrades.py
@@ -1,0 +1,73 @@
+# -*- coding: utf-8 -*-
+from plone import api
+from sc.embedder.config import PROFILE
+from sc.embedder.testing import INTEGRATION_TESTING
+
+import unittest
+
+
+class UpgradeTestCaseBase(unittest.TestCase):
+
+    layer = INTEGRATION_TESTING
+
+    def setUp(self, from_version, to_version):
+        self.portal = self.layer['portal']
+        self.setup = self.portal['portal_setup']
+        self.from_version = from_version
+        self.to_version = to_version
+
+    def _get_upgrade_step(self, title):
+        """Get one of the upgrade steps.
+
+        :param title: [required] the title used to register the upgrade step
+        :type obj: str
+        """
+        self.setup.setLastVersionForProfile(PROFILE, self.from_version)
+        upgrades = self.setup.listUpgrades(PROFILE)
+        steps = [s for s in upgrades[0] if s['title'] == title]
+        return steps[0] if steps else None
+
+    def _do_upgrade_step(self, step):
+        """Execute an upgrade step.
+
+        :param step: [required] the step we want to run
+        :type step: str
+        """
+        request = self.layer['request']
+        request.form['profile_id'] = PROFILE
+        request.form['upgrades'] = [step['id']]
+        self.setup.manage_doUpgrades(request=request)
+
+    def _how_many_upgrades_to_do(self):
+        self.setup.setLastVersionForProfile(PROFILE, self.from_version)
+        upgrades = self.setup.listUpgrades(PROFILE)
+        assert len(upgrades) > 0
+        return len(upgrades[0])
+
+
+class Upgrade1000to1001TestCase(UpgradeTestCaseBase):
+
+    def setUp(self):
+        UpgradeTestCaseBase.setUp(self, u'1000', u'1001')
+
+    def test_upgrade_to_1001_registrations(self):
+        version = self.setup.getLastVersionForProfile(PROFILE)[0]
+        self.assertGreaterEqual(int(version), int(self.to_version))
+        self.assertEqual(self._how_many_upgrades_to_do(), 1)
+
+    def test_add_embedder_tile(self):
+        # check if the upgrade step is registered
+        title = u'Add embedder tile'
+        step = self._get_upgrade_step(title)
+        self.assertIsNotNone(step)
+
+        # simulate state on previous version
+        from sc.embedder.Extensions.Install import remove_tile
+        remove_tile(api.portal.get())
+        tiles = api.portal.get_registry_record('plone.app.tiles')
+        self.assertNotIn(u'sc.embedder', tiles)
+
+        # run the upgrade step to validate the update
+        self._do_upgrade_step(step)
+        tiles = api.portal.get_registry_record('plone.app.tiles')
+        self.assertIn(u'sc.embedder', tiles)